+++ xmhf/src/xmhf-core/xmhf-runtime/xmhf-xcphandler/xcph-interface.c
@@ -65,6 +65,6 @@ u8 * xmhf_xcphandler_get_idt_start(void){
 
 
 //EMHF exception handler hub
-void xmhf_xcphandler_hub(u32 vector, struct regs *r){
+void xmhf_xcphandler_hub(uintptr_t vector, struct regs *r){
 	xmhf_xcphandler_arch_hub(vector, r);
 }
